feat(www): refactor ModelMatrix component and introduce new subcomponents
- Refactored the ModelMatrix component to improve structure and readability by breaking it down into smaller subcomponents: GroupedModels, LastUpdated, ProviderSidebar, ModelCard, and SearchControls.
- Enhanced model data handling with utility functions for building models from the catalog, filtering, and grouping by provider.
- Updated the types for better type safety and clarity across the component.
- Improved search functionality and provider selection interface for a better user experience.
